style: Rename component Location -> Map
